# huge memory optimizations