--- a/lark/parsers/grammar_analysis.py
+++ b/lark/parsers/grammar_analysis.py
@@ -3,8 +3,12 @@ from ..utils import bfs, fzset, classify
 from ..common import GrammarError, is_terminal
 from ..grammar import Rule
 
+import weakref
+
 
 class RulePtr(object):
+    __slots__ = ('rule', 'index', 'next')
+
     def __init__(self, rule, index):
         assert isinstance(rule, Rule)
         assert index <= len(rule.expansion)
@@ -105,6 +109,7 @@ def calculate_sets(rules):
 class GrammarAnalyzer(object):
     def __init__(self, parser_conf, debug=False):
         self.debug = debug
+        self.parser_conf = parser_conf
 
         rules = parser_conf.rules + [Rule('$root', [parser_conf.start, '$END'])]
         self.rules_by_origin = classify(rules, lambda r: r.origin)
@@ -134,7 +139,8 @@ class GrammarAnalyzer(object):
                     if not is_terminal(new_r):
                         yield new_r
 
-        _ = list(bfs([rule], _expand_rule))
+        for _ in bfs([rule], _expand_rule):
+            pass
 
         return fzset(init_ptrs)