First of all apologies for the big changeset, I'll comment on the PR to explain some parts better.

Here are some bench numbers:

### Baseline (master)

    test_python_lib (22 files), time: 0.936244010925 secs
    Partition of a set of 275823 objects. Total size = 46740096 bytes.
    Index  Count   %     Size   % Cumulative  % Kind (class / dict of class)
        0  99403  36 27832840  60  27832840  60 dict of lark.parsers.grammar_analysis.RulePtr
        1  99403  36  6361792  14  34194632  73 lark.parsers.grammar_analysis.RulePtr
        2   1685   1  2937080   6  37131712  79 dict (no owner)
        3  30415  11  2263704   5  39395416  84 tuple
        4   6514   2  1563360   3  40958776  88 lark.utils.fzset
        5  16075   6  1268840   3  42227616  90 str
        6    613   0   590968   1  42818584  92 dict of lark.lexer.Token
        7   2618   1   366672   1  43185256  92 list
        8    407   0   364304   1  43549560  93 type


### Optimizations

 - Speed: -13%
 - Memory: -565%
```
    test_python_lib (22 files), time: 0.826800107956 secs
    Partition of a set of 61184 objects. Total size = 8259976 bytes.
    Index  Count   %     Size   % Cumulative  % Kind (class / dict of class)
        0  23976  39  1801152  22   1801152  22 tuple
        1   1012   2  1667296  20   3468448  42 dict (no owner)
        2  16202  26  1278104  15   4746552  57 str
        3    408   1   365848   4   5112400  62 type
        4    112   0   347776   4   5460176  66 dict of module
        5   2470   4   345776   4   5805952  70 list
        6    408   1   337728   4   6143680  74 dict of type
        7   2314   4   296192   4   6439872  78 types.CodeType
        8   2099   3   251880   3   6691752  81 function
        9    169   0   154072   2   6845824  83 dict of class
```
### ArrayParseTable (no cache)

 - Speed: +56%
 - Memory: -739%
```
    test_python_lib (22 files), time: 1.65054392815 secs
    Partition of a set of 48642 objects. Total size = 6324520 bytes.
    Index  Count   %     Size   % Cumulative  % Kind (class / dict of class)
        0  16202  33  1278104  20   1278104  20 str
        1  10854  22   856384  14   2134488  34 tuple
        2    425   1   562904   9   2697392  43 dict (no owner)
        3    408   1   365848   6   3063240  48 type
        4   2471   5   348920   6   3412160  54 list
```
### ArrayParseTable (cached)

 - Speed: +5%
 - Memory: -669%
```
    test_python_lib (22 files), time: 0.975975990295 secs
    Partition of a set of 53713 objects. Total size = 6978824 bytes.
    Index  Count   %     Size   % Cumulative  % Kind (class / dict of class)
        0  16174  30  1276744  18   1276744  18 str
        1  14268  27  1102200  16   2378944  34 tuple
        2   1013   2   946424  14   3325368  48 dict (no owner)
        3    408   1   365848   5   3691216  53 type
        4   2471   5   348920   5   4040136  58 list
```