@drslump 2018-04-07
Improved the `ArrayParseTable` performance (and removed the cache variant), while slower it could be a nice alternative for memory constrained systems, specially if it's supported on the standalone generator.

- Speed: +64%
- Memory: -738%

```
test_python_lib (22 files), time: 1.42830491066 secs
Partition of a set of 48672 objects. Total size = 6334136 bytes.
 Index  Count   %     Size   % Cumulative  % Kind (class / dict of class)
     0  16213  33  1278856  20   1278856  20 str
     1  10864  22   857008  14   2135864  34 tuple
     2    425   1   562904   9   2698768  43 dict (no owner)
     3    408   1   365888   6   3064656  48 type
     4   2471   5   348920   6   3413576  54 list
     5    112   0   347776   5   3761352  59 dict of module
     6    408   1   337728   5   4099080  65 dict of type
     7   2317   5   296576   5   4395656  69 types.CodeType
     8   2101   4   252120   4   4647776  73 function
     9    169   0   154072   2   4801848  76 dict of class
```