@erezsh 2018-04-07
I'm glad you submitted it. It contains several good fixes.

You're right about the analysis instance. I kept it for debugging purposes, but that's not a good enough reason to double the memory usage. I'll take care of it.

The problem with the Tree class is that some users use it to store their own information. But Lark accepts a tree_class parameter, so it should be able to accept a slotted tree (or I can make sure that it will).

`ArrayParseTable` is a pretty cool hack. I'm still not sure why it saves so much memory (didn't test it for myself though, just taking your word for it). But between adding 50% to the run-time and limiting the size of the parse table, I'm reluctant to add it as an official feature. But maybe I can adopt the `parsetable_class` option, and I'll add it to the documentation as a "recipe".

Lastly, if I keep pushing patches on my own, you won't get any "official" credit for your helpful work (I'll mention you ofc, but it won't count in github as a contribution). We can get around that, If you'd like. I can create a pull-request on your fork with my changes, and you can open a new PR. Let me know if you want to do that.