@drslump 2018-04-07
No problem in issuing the commits yourself, thanks for that!

The size difference with ArrayParseTable is because a simple int in Python takes around 24 bytes while inside the array it's just 2 bytes, so the bigger the tables the bigger the savings.

Offtopic, I just noticed the standalone parser is licensed under GPL, may I ask if there is a reasoning behind it? I was planing on generating a parser for a project but I can't use GPL there.